+++ EU4toV2/CMakeLists.txt
@@ -36,6 +36,7 @@ file(GLOB MAPPER_COLONIALTAGS_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/ColonialTag
 file(GLOB MAPPER_COLONIALFLAGS_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/ColonialFlags/*.cpp")
 file(GLOB MAPPER_CULTUREMAPPER_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/CultureMapper/*.cpp")
 file(GLOB MAPPER_FLAGCOLORS_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/FlagColors/*.cpp")
+file(GLOB MAPPER_GOVERNMENTMAPPER_SOURCES "${PROJECT_SOURCE_DIR}/Mappers/GovernmentMapper/*.cpp")
 file(GLOB EU4_WORLD_SOURCES "${PROJECT_SOURCE_DIR}/EU4World/*.cpp")
 file(GLOB EU4_COLONIALREGIONS_SOURCES "${PROJECT_SOURCE_DIR}/EU4World/ColonialRegions/*.cpp")
 file(GLOB EU4_HISTORY_SOURCES "${PROJECT_SOURCE_DIR}/EU4World/History/*.cpp")
@@ -85,6 +86,7 @@ add_executable(EU4ToVic2
 	${MAPPER_COLONIALFLAGS_SOURCES}
 	${MAPPER_CULTUREMAPPER_SOURCES}
 	${MAPPER_FLAGCOLORS_SOURCES}
+	${MAPPER_GOVERNMENTMAPPER_SOURCES}
 	${EU4_DIPLOMACY_SOURCES}
 	${EU4_RELATIONS_SOURCES}
 	${EU4_NATIONMERGER_SOURCES}
