+++ EU4toV2/EU4ToV2.vcxproj
@@ -251,7 +251,7 @@
     <ClCompile Include="Source\Mappers\Geography\ClimateMapper.cpp" />
     <ClCompile Include="Source\Mappers\Geography\Continents.cpp" />
     <ClCompile Include="Source\Mappers\Geography\TerrainDataMapper.cpp" />
-    <ClCompile Include="Source\Mappers\GovermentMapper.cpp" />
+    <ClCompile Include="Source\Mappers\GovernmentMapper\GovernmentMapper.cpp" />
     <ClCompile Include="Source\Mappers\Ideas\IdeaEffectMapper.cpp" />
     <ClCompile Include="Source\Mappers\Ideas\IdeaEffects.cpp" />
     <ClCompile Include="Source\Mappers\Ideas\TechGroups.cpp" />
@@ -402,7 +402,7 @@
     <ClInclude Include="Source\Mappers\Geography\ClimateMapper.h" />
     <ClInclude Include="Source\Mappers\Geography\Continents.h" />
     <ClInclude Include="Source\Mappers\Geography\TerrainDataMapper.h" />
-    <ClInclude Include="Source\Mappers\GovernmentMapper.h" />
+    <ClInclude Include="Source\Mappers\GovernmentMapper\GovernmentMapper.h" />
     <ClInclude Include="Source\Mappers\Ideas\IdeaEffectMapper.h" />
     <ClInclude Include="Source\Mappers\Ideas\IdeaEffects.h" />
     <ClInclude Include="Source\Mappers\Ideas\TechGroups.h" />
