+++ EU4toV2/Source/Mappers/GovernmentMapper/GovernmentMapper.h
@@ -4,16 +4,17 @@
 #include <map>
 #include "newParser.h"
 
-
 namespace mappers
 {
 	class GovernmentMapper : commonItems::parser
 	{
 	public:
+		GovernmentMapper();
 		GovernmentMapper(std::istream& theStream);
 		std::string matchGovernment(const std::string& sourceGovernment) const;
 
 	private:
+		void registerKeys();
 		std::map<std::string, std::string> governmentMap;
 	};
 };
